:recycle: Created mappings module from spans_to_ functions of Regen
